api/requirements.txt
requests==2.27.0
PyYAML==6.0